Add user entity
Migration populates user table with distinct owner values from short_urls table.
Remove user security class
