+++ config/packages/security.yaml
@@ -1,7 +1,10 @@
 security:
     providers:
-        short_user_provider:
-            id: App\Security\UserProvider
+        # used to reload user from session & other features (e.g. switch_user)
+        app_user_provider:
+            entity:
+                class: App\Entity\User
+                property: id
 
     firewalls:
         dev:
