--- src/Security/User.php
@@ -1,63 +1,0 @@
-<?php
-
-namespace App\Security;
-
-use Symfony\Component\Security\Core\User\UserInterface;
-
-class User implements UserInterface
-{
-    private $userid;
-    private $roles = [];
-
-    public function __construct(string $username, bool $admin = false)
-    {
-        $this->userid = $username;
-        $this->roles[] = $admin ? 'ROLE_ADMIN' : 'ROLE_USER';
-    }
-
-    /**
-     * A visual identifier that represents this user.
-     *
-     * @see UserInterface
-     */
-    public function getUsername(): string
-    {
-        return (string) $this->userid;
-    }
-
-    /**
-     * @see UserInterface
-     */
-    public function getRoles(): array
-    {
-        $roles = $this->roles;
-
-        // guarantees that a user always has at least one role for security
-        if (empty($roles)) {
-            $roles[] = 'ROLE_USER';
-        }
-
-        return array_unique($roles);
-    }
-
-    /**
-     * @see UserInterface
-     */
-    public function getPassword()
-    {
-    }
-
-    /**
-     * @see UserInterface
-     */
-    public function getSalt()
-    {
-    }
-
-    /**
-     * @see UserInterface
-     */
-    public function eraseCredentials()
-    {
-    }
-}
